admin_penel/requirements.txt
django==3.2
python-dotenv==0.20.0
psycopg2-binary==2.9.3
django_split_settings==1.1.0
django-debug-toolbar==3.4.0
gunicorn==20.1.0